Vertical-Banking/requirements.txt
pandas
crewai
crewai_tools
load_dotenv
langchain-huggingface